pyproject.toml
[project]
name = "compliance-collector"
version = "0.1.0"
description = "Open-source compliance evidence auto-collector for Microsoft 365 and Azure."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Syed Aun Muhammad" }]
keywords = ["microsoft365", "azure", "compliance", "soc2", "iso27001", "cis", "audit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "msgraph-sdk>=1.5.0",
    "azure-identity>=1.16.0",
    "ruamel.yaml>=0.18.6",
    "jinja2>=3.1.4",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
compliance-collector = "compliance_collector.cli:app"

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/compliance-collector"
Issues = "https://github.com/YOUR_USERNAME/compliance-collector/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/compliance_collector"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=compliance_collector --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
